backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
aiomysql==0.2.0
sqlalchemy==2.0.23
pymysql==1.1.0
cryptography==41.0.7

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
bcrypt==4.1.1
argon2-cffi==23.1.0

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Security Headers and Rate Limiting
slowapi==0.1.9
secure==0.3.0

# # CORS
# python-cors==1.0.0

# Utilities
httpx==0.25.2
python-dateutil==2.8.2

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1

